stt_project/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Deprecated==1.2.18
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Limiter==3.12
fsspec==2025.5.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==1.24.4
openai-whisper==20240930
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.1
regex==2024.11.6
requests==2.32.4
rich==13.9.4
semantic-version==2.10.0
setuptools-rust==1.11.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
